Fix tailwindcss processing
The Blist tailwindcss/postcss processing was creating a parallel
CSS processing pipeline to the one setup at the theme level.  This lead
to misleading successes which were exposed when the config.toml file
was renamed to hugo.toml.

While it the Blist configuration has what appears to be some
anachronisms from tailwindcss v2, it works.  So rather than try to get
my purely tailwind v3 setup perfected, remove that config altogether and
let the Blist configuration with postcss handle it all.