Fix markdown ignoring newlines
This will make the raw text files harder to read, but is necessary for
markdown in other places to render correctly.  An example of this is
the address field in hugo.toml; without hardWrap the address is
arbitrarily wrapped in inappropriate ways.

So live with this for now and if it becomes a problem look for a better
solution (like is individual page config of this setting possible?)