Fix service-card.html issue rendering tag lists
When the service-card partial is used to render the tag lists, it isn't
passed a thumbnail.  The service-card partial had no case where it would
render anything (of note) when there was no thumbnail and so the tags
list didn't render any services related tags at all.

Make a simple text listing for tags in this list context.  Note that
we'll likely want to style this better than it is at present.