Add section navigation to single page layout
This includes making it configurable by type and by page.

Note that there is a known issue in that the default of "false" fails to
produce a correct result when no configuration exists at the global or
page levels.  Didn't troubleshoot too much and made sure there's at
least a global config for both page types.