Add contact & contact thank you forms
Add the contact form to the main menu as well.

Finally, remove the "Tags" menu item by commenting it out of hugo.toml.

We want to hold on to the tags entry in hugo.toml, because if we start
blogging here we'll use the tags feature (and need to remember how it
should look); but until that time having it is just a useless
distraction.